Create time series mini bar chart
