+++ packages/app/src/components-styled/time-series-chart/components/benchmark.tsx
@@ -1,27 +1,44 @@
 import { colors } from '~/style/theme';
 import { Group } from '@visx/group';
 import { Line } from '@visx/shape';
+import { Ref } from 'react';
 
 interface BenchmarkProps {
   value: number;
   label?: string;
   top: number;
   width: number;
+  /**
+   * This ref can be used for measuring the width of the Y-axis to automagically
+   * calculate a left-padding.
+   */
+  textRef?: Ref<SVGGElement>;
+
+  xRangePadding?: number;
 }
 
 const color = colors.data.benchmark;
 
-export function Benchmark({ top, value, label, width }: BenchmarkProps) {
+export function Benchmark({
+  top,
+  value,
+  label,
+  width,
+  textRef,
+  xRangePadding = 0,
+}: BenchmarkProps) {
   return (
     <Group top={top}>
-      <text fontSize="14px" dy={-4} fill={color}>
-        {label ? `${label}: ${value}` : value}
-      </text>
+      <g ref={textRef}>
+        <text fontSize="14px" dx={-xRangePadding} dy={-4} fill={color}>
+          {label ? `${label}: ${value}` : value}
+        </text>
+      </g>
       <Line
         stroke={color}
         strokeDasharray="4,3"
-        from={{ x: 0, y: 0 }}
-        to={{ x: width, y: 0 }}
+        from={{ x: 0 - xRangePadding, y: 0 }}
+        to={{ x: width + xRangePadding, y: 0 }}
       />
     </Group>
   );
