+++ packages/app/src/components-styled/time-series-chart/index.ts
@@ -1,2 +1,2 @@
 export * from './time-series-chart';
-export * from './mini-time-series-chart';
+export * from './time-series-mini-bar-chart';
